added option --embed-bitmaps to embed bitmap data into the SVG files instead of adding file references

+++ src/XMLNode.cpp
@@ -375,7 +375,8 @@ ostream& XMLElement::write (ostream &os) const {
 		if (attrib.name.front() != '@')
 			os << attrib.name << "='" << attrib.value << '\'';
 		else {
-			os << attrib.name.substr(1) << "='";
+			bool keep = (attrib.name.size() > 1 && attrib.name[1] == '@');
+			os << attrib.name.substr(keep ? 2 : 1) << "='";
 			auto pos = attrib.value.find("base64,");
 			if (pos == string::npos)
 				os << attrib.value;
@@ -387,7 +388,7 @@ ostream& XMLElement::write (ostream &os) const {
 					os << '\n';
 					util::base64_copy(ifs, os, 200);
 					ifs.close();
-					if (!KEEP_ENCODED_FILES)
+					if (!KEEP_ENCODED_FILES && !keep)
 						FileSystem::remove(fname);
 				}
 			}
